Ver0.05
EnemyAnimationState

+++ Assets/Scripts/Interfaces/Health.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface Health
 {
-    float health { get; set; }
+    int health { get; set; }
 }
